refactor(traverse): clean up code and remove useless comments

+++ storage.py
@@ -20,7 +20,7 @@ class Storage:
         self.conn = conn
         if self.drop:
             self.conn.executescript(queries.DROP_TABLES)
-        self.conn.execute(queries.MAKE_TABLES)
+        self.conn.executescript(queries.MAKE_TABLES)
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
@@ -28,49 +28,49 @@ class Storage:
             print("ex")
             # log exceptions
             pass
-        # TODO autocommit l0l
-        self.conn.commit()
         self.conn.close()
+        return False
 
     def save_file(self, filename, parent_dir, permissions, sha256):
         file_record = self.conn.execute(
-            "SELECT * from file where dir_id = ? and filename = ?", (parent_dir,
-                                                                     filename)).fetchone()
+            "SELECT * from file where dir_id = ? and filename = ?",
+            (parent_dir, filename)).fetchone()
         if not file_record:
             # insert
             self.conn.execute(queries.INSERT_FILE, (
                 parent_dir, filename, datetime.now(), permissions, sha256))
         else:
             self.conn.execute(
-                "UPDATE file set last_modified = ?, permissions = ?, sha256 = ? where id = ?",
+                "UPDATE file set last_modified = ?, permissions = ?, sha256 = ?"
+                " where id = ?",
                 (datetime.now(), permissions, sha256, file_record['id']))
 
     def save_dir(self, dir_path, parent_dir):
         return self.conn.execute(queries.INSERT_DIR, (parent_dir, dir_path))
 
-    def old_dirs(self, root_id):
+    def fetch_subdirs(self, dir_id):
         # select all dirs
         dir_cache = {}
         result = self.conn.execute(
-            "SELECT id, parent_dir_id, dir_path from directory where parent_dir_id = ?",
-            (root_id,))
+            "SELECT id, parent_dir_id, dir_path from directory "
+            "where parent_dir_id = ?", (dir_id,))
 
-        for id, parent_dir_id, dir_path in result:
-            dir_cache[dir_path] = id
+        for dir_id, parent_dir_id, dir_path in result:
+            dir_cache[dir_path] = dir_id
         return dir_cache
 
-    def old_files(self, root):
+    def fetch_files(self, dir_path):
         result = self.conn.execute(
             "SELECT file.filename,file.id from file join directory "
-            "on dir_id=directory.id where directory.dir_path = ?",
-            (root,))
+            "on dir_id=directory.id "
+            "where directory.dir_path = ?", (dir_path,))
         file_cache = {}
         for filename, file_id in result:
             file_cache[filename] = file_id
         return file_cache
 
-    def drop_files(self, file_ids):
-        if file_ids:
+    def drop_files(self, file_ids: Sequence):
+        if file_ids and len(file_ids) != 0:
             self.conn.execute(f"DELETE FROM file where id in ("
                               f"{self._plug_params(file_ids)})", file_ids)
